apps/backend/requirements.txt
# TMDB Pipeline Requirements
requests>=2.28.0
urllib3>=1.26.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# API Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
thefuzz>=0.20.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.24.0
